sovereign-advisor-v2/requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
langchain
langchain-community
langchain-text-splitters
pypdf
python-frontmatter
python-docx
qdrant-client
fastembed
